requirements.txt
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
fastapi==0.61.2
h11==0.11.0
idna==2.10
pydantic==1.7.2
python-dotenv==0.15.0
requests==2.25.0
SQLAlchemy==1.3.20
starlette==0.13.6
urllib3==1.26.2
uvicorn==0.12.2